+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Enhancement
 - [#121] Improve positioning of "GM Threat" and "Party Momentum" boxes
 - [#184] Show tooltips for Armor Qualities
+- [#185] Provide a method of indicating an NPC weapon is an escalation option
 
 ### Chores
 - [#186] Migrate momentum tracker application to ApplicationV2